Fix through2 transform and flush functions
The tests were assuming the code was synchronous.  Once
```done()``` was introduced a few tests failed.

As part of a previous commit #fd8e1dc79b, the library "through"
was replaced to "through2".  In order for the parsed value to be
returned as a string the ```{ objectMode: true }``` options object
is needed.